# docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact: {}
  description: 荆楚花牌(花好月圆)游戏服务端 API
  title: 荆楚花牌 API
  version: "1.0"
paths:
...
  /rooms:
    get:
      description: 获取所有游戏房间列表
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"rooms\": [] }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 列出所有房间
      tags:
      - 房间
    post:
      consumes:
      - application/json
      description: 创建新游戏房间，创建者自动加入
      parameters:
      - description: 房间信息
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"room_id\": \"string\", \"room\": {} }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 创建房间
      tags:
      - 房间
  /rooms/{id}:
    get:
      description: 根据房间 ID 查询房间详情
      parameters:
      - description: 房间 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 房间信息
          schema:
            type: object
        "404":
          description: '{ \"error\": \"room not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 查询房间
      tags:
      - 房间
  /rooms/{id}/ai:
    post:
      description: 向房间添加一个AI玩家（AI自动准备）
      parameters:
      - description: 房间 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: AI 玩家信息
          schema:
            type: object
        "400":
          description: '{ \"error\": \"room is full\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"room not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 添加AI玩家
      tags:
      - 房间
  /rooms/{id}/join:
    post:
      description: 加入指定房间
      parameters:
      - description: 房间 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 房间信息
          schema:
            type: object
        "400":
          description: '{ \"error\": \"room is full\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 加入房间
      tags:
      - 房间
  /rooms/{id}/leave:
    post:
      description: 离开指定房间
      parameters:
      - description: 房间 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"ok\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"string\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 离开房间
      tags:
      - 房间
...
  /rooms/{id}/start:
    post:
      description: 房主开始游戏，需要房间已满且所有玩家已准备
      parameters:
      - description: 房间 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ \"status\": \"started\", \"room_id\": \"string\" }'
          schema:
            type: object
        "400":
          description: '{ \"error\": \"room is not full\" }'
          schema:
            type: object
        "404":
          description: '{ \"error\": \"room not found\" }'
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 开始游戏
      tags:
      - 房间
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"